First Quarter Report
Work still in progress

+++ DTO/AdminReport.cs
@@ -9,22 +9,22 @@
         public double ASalandWages { get; set; }
         public double AEquipment { get; set; }
         public string RegionName { get; set; }
-        public int AEmpBenefits { get; set; }
-        public int AEmpTravel { get; set; }
-        public int AEmpTraining { get; set; }
-        public int AOfficeRent { get; set; }
-        public int AOfficeUtilities { get; set; }
-        public int AFacilityIns { get; set; }
-        public int AOfficeSupplies { get; set; }
-        public int AOfficeCommunications { get; set; }
-        public int AOfficeMaint { get; set; }
-        public int AConsulting { get; set; }
-        public int AJanitorServices { get; set; }
-        public int ADepreciation { get; set; }
-        public int ATechSupport { get; set; }
-        public int ASecurityServices { get; set; }
-        public int AOther { get; set; }
-        public int AOther2 { get; set; }
-        public int AOther3 { get; set; }
+        public double AEmpBenefits { get; set; }
+        public double AEmpTravel { get; set; }
+        public double AEmpTraining { get; set; }
+        public double AOfficeRent { get; set; }
+        public double AOfficeUtilities { get; set; }
+        public double AFacilityIns { get; set; }
+        public double AOfficeSupplies { get; set; }
+        public double AOfficeCommunications { get; set; }
+        public double AOfficeMaint { get; set; }
+        public double AConsulting { get; set; }
+        public double AJanitorServices { get; set; }
+        public double ADepreciation { get; set; }
+        public double ATechSupport { get; set; }
+        public double ASecurityServices { get; set; }
+        public double AOther { get; set; }
+        public double AOther2 { get; set; }
+        public double AOther3 { get; set; }
     }
 }
